IE8 32-bit Support
Ошибка выдачи сообщения об ошибке

Более подробно

http://stackoverflow.com/questions/690251/what-happened-to-console-log-in-ie8

console.log is only available after you have opened the Developer Tools
(F12 to toggle it open and closed). Funny thing is that after you've
opened it, you can close it, then still post to it via console.log
calls, and those will be seen when you reopen it. I'm thinking that is a
bug of sorts, and may be fixed, but we shall see.